separating functions on search.py
